+++ fpdf/image_parsing.py
@@ -20,6 +20,11 @@ try:
 except ImportError:
     Image = None
 
+try:
+    import numpy as np
+except ImportError:
+    np = None
+
 from .errors import FPDFException
 from .image_datastructures import ImageCache, RasterImageInfo, VectorImageInfo
 from .svg import SVGObject
@@ -614,6 +619,8 @@ def pack_codes_into_bytes(codes):
     bits_in_buffer = 0
     output = bytearray()
 
+    if np is not None:
+        codes = np.array(codes, dtype=np.uint32)
     for code in codes:
         buffer = (buffer << bits_per_code) | code
         bits_in_buffer += bits_per_code